Add 32 bit support (int/float) support to audioinput example.

+++ examples/audioinput/audioinput.h
@@ -69,7 +69,7 @@ public:
 
 private:
     const QAudioFormat m_format;
-    quint16 m_maxAmplitude;
+    quint32 m_maxAmplitude;
     qreal m_level; // 0.0 <= m_level <= 1.0
 
 signals:
